Adds the api scripts
Initial commit of the api, its commented pretty decently. The two files in core are models, that is where the sql command is located and will serve as the model for any end point relating to their respective tables. Read contact is a tester end point. Database in the config folder holds all of the database info and well as the function to establish a connection to it, so that it is only listed once just in case anything changes.